UI #90: 재사용 가능한 컴포넌트 common 분리 및 적용
