+++ fe/src/constant/index.ts
@@ -22,6 +22,7 @@ export const MESSAGE = {
     "북크북크 북클럽에 오신걸 환영합니다! \n 북클럽에 등록중이니 잠시만 기다려주세요!",
   BOOK_CLUB_JOIN_ERROR: `북클럽 가입에 실패했어요. \n 나중에 다시 시도해주세요.`,
   NOT_FOUND: `요청하신 페이지를 찾을 수 없어요. \n 다른 페이지로 이동해주세요.`,
+  NEW_GATHERING_ERROR: `모임 생성에 실패했어요. \n 나중에 다시 시도해주세요.`,
 };
 
 export const PAGE_TITLE = {
